batman-adv: Convert existing sysfs options to netlink
The current netlink options are only available via sysfs. While the new
netlink interface could only be used for new options, it is easier for the
userspace to have a unified interface which works for new and old options.

Most options in sysfs are already simple values and their sysfs filenames
can be used as key for the new netlink configuration interface. But the
multi-value files like gw_bandwidth and isolation_mark can be split to
better fit in their batadv internal representation:

* isolation_mark
   - isolation_mask
   - isolation_mark
* gw_bandwidth
  - gw_bandwidth_down
  - gw_bandwidth_up

TODO add more description here.

Cc: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: Sven Eckelmann <sven@narfation.org>